Revert "Move SpecialFunctions to an extension (#564)"
This reverts commit cf2707323214742fd934f248ad0199b320912d0a.
